+++ packages/core/CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning].
 - Add support for mangling JavaScript embedded in HTML.
 - Fix mangling CSS values when `!important` is used.
 - Fix mangling CSS properties, values, and selectors when comments are used.
+- Prevent mangling in CSS, HTML, and JavaScript comments and strings.
 
 ## [0.1.20] - 2021-05-13